change PanelViewFilterVariantCallingQuality.ts
